issue-4875: moved channels state to separate TPartitionChannelsState class
