+++ cloud/blockstore/libs/storage/partition_common/ut/ya.make
@@ -8,6 +8,7 @@ SRCS(
     actor_loadfreshblobs_ut.cpp
     drain_actor_companion_ut.cpp
     long_running_operation_companion_ut.cpp
+    part_channels_state_ut.cpp
 )
 
 PEERDIR(
